requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
alembic==1.13.0
sqlalchemy==2.0.23
pymongo-migrate==1.0.0
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0.post1